Simplify declaration of aparams.
This is just a consequence of me reading code and seeing things that can be tidied a little. If this is not helpful that please say and we can just throw this away completely.
By using literal maps, and moving away from strings a bit, the code is a bit clearer to read, and it caught a bug too (ASPREFERENCE was defined twice).