Refactor robot instructions.
The main goal here has been to move away from so much reliance of strings, which tend to be highly fragile and prone to mis-typing. In addition, by using anonymous embedding of structs, we can start introducing default behaviours, and reduce the amount of boilerplate, and reduce some of the level of casting going on (for example, see how I've moved down the merging of MOV instructions with ASP/DSP/MIX).
The tests pass.

I've chosen to stop at the point where the instruction parameters start coming out of the LHPolicy because it's not clear to me how to unify them, and this PR is big enough already. There's a good chance I've made some mistaken changes somewhere which the tests won't catch, and sadly this PR is fairly big so it may be a bit tricky to review.