Remove promises.
The promises implementation was basically not being used - everything being issued as a promise was in fact a constant value, and so the entire promises mechanism was about an extra 1000 lines of code creating confusion and complexity. It can always be brought back when it's needed - and simply too - as can be seen from the diff here, the changes to intrinsics are trivial so can easily be undone.

local tests pass, and the fuller antha-runner/antha-tester tests pass too.