In codegen/findBestMoveDevice, minC was set to -1 which is potentially a bug if any of the moveCosts are < 0. Similarly, partition_tree/runExact has the same bug.
Typically, if you are trying to find a min, you need to start your accumulator with the maximum possible value. This is what I've changed it to.
To do this required moving a bunch of vars from int to int64 so that we can use math.MaxInt64 and have certainty about the width of the var.

I did try moving to a uint64 - this would be my preference, especially on the interface containing MoveCost. In fact, in tree_partition there is even a comment that EdgeWeight cannot be < 0. However, sadly in sssp, a negative weight is used in order to force calculation of the longest path, rather than the shortest path, and reworking all of that would have been painful. Hence I decided to compromise on int64.

Local antha-lang tests pass, and antha-tester tests pass as before.