Unify LHInstruction.Add{Product,Result} into .AddOutput.
Not sure what to do about AddComponent: I'm happy to change it to AddInput, but there are 3 elements that I can find that use it (Transformation, Setup_Input_Plate, Transformation_Multi). Just fearful of changing a public API that may be used by 3rd party elements - how do we normally handle this case?