Be more accurate about labelling a file as "not json".
Previously, we were trying to unmarshal into a Bundle object. If that failed, we claimed it wasn't a valid json file. But this is not correct - an array is a valid json file, but because a bundle is an object, it won't unmarshal.
So if we unmarshal to json.RawMessage first then it just checks it's valid json.

This is related to antha-runner PR370