Refactor user layout preferences (#935)
Layout preferences determine which types of objects are to be placed in which deck location, and in which order.

There are two sources of layout preferences: the device plugin and the user. Previously, these sources are merged according to complex, device specific rules which are hardcoded into antha, meaning that devices can't be added without adding to this hardcode.

This change introduces a simple device independent method for merging the user positions with the driver positions:

if no user preferences are provided, the plugin provided list is used
otherwise, user preferences are accepted in order if they are also specified in the device's list, otherwise they are ignored.
This allows the device to specify which locations can possibly hold a specific class of labware, and the user to specify which locations should hold a specific class of labware. If the user specifies a position which doesn't exist, it is silently ignored, such that specifying "position_1" to a hamilton driver has no effect.