LHWell .Protect is not used.
Context: Plates seem odd. They're in wtype and thus seem available to elements, but are sorta managed through sampletracker which is in microArch...
I was trying to figure out if we can drop all pointers to Liquids and just pass around maps from ids to objects, which would at least move us away from the weirdness of Dup with IDs (I found some elements that call Dup()...)
Which then made me think of Plates, which contain Wells, which contain Liquids... so I was looking at Wells and came across the protect stuff.