Fix/volume error250119 (#957)
* Fix: excessive volumes in user plates (Part 1)

Summary:

Users have reported autogenerated input plates requesting volumes
above the maximum allowed for the well type in question.

Owing to some complexities in the process with the handling of
residual and carry volumes this is tolerated for small
values but the values in question were much too large to be a
consequence of this.

Investigation of this found that in fact there was a fairly serious
error in which two multichannel transfers successively could generate
an extra channel unintentionally. e.g. a 7 channel transfer followed
by a 6 channel transfer would create two 7 channel transfers.

This change modifies the logic in cases like this to account for the
first transfer exhausting one or more wells.

Test Plan:

Antha-tester, manual validation against original bug.

Reviewers:

Haydn King

Subscribers:

Nuno Leitao

Projects:

Antha language

* Prevent excessive autoallocated volumes (part 2)

Summary:

Carry and residual volumes are necessary to prevent overestimation of
liquid volumes. However they occasionally lead to slightly excessive
volumes being requested.

This change simply ensures that volumes reported to users are at most
the maximum allowed per well.

This is undesirable, but tolerable since we believe the carry volumes
are overestimated and decrease as wells are emptied, so that in the
cases this is required the actual volume taken from the well is lower
than requested by a similar amount to the excess added for carry
volumes.

Test Plan:

Antha-tester, manual testing

Reviewers:

Haydn King

Subscribers:

Projects:

Antha language

* Prevent deleting volumes creating discontiguities

Summary:

Test Plan:

Reviewers:

Subscribers:

Projects:

* Shut up gometalinter

Summary:

Test Plan:

Reviewers:

Subscribers:

Projects:

* transferparams.Zero --> transferparams.IsZero

Summary:

Consistency fix

Test Plan:

Reviewers:

Haydn King

Subscribers:

Projects:

Antha-language

* RemoveInitialBlanks now does what it says

Summary:

the old function still exists but is now called RemoveBlanks

Test Plan:

Reviewers:

Haydn King

Subscribers:

Projects:

Antha language

* Noncontiguous migrated to volume set method

Summary:

Additionally the names of the two variables used in the check were
made more descriptive of their actual meaning

Test Plan:

Reviewers:

Haydn King

Subscribers:

Projects:

Antha-language

* Catch autoallocated volumes exceeding tolerance

Summary:

Volume autoallocation is tricky because it has to include some aspect of
simulation since the actual volume required in a well is a function of
both the amount taken and the number and kind of transfers used.

As a consequence of the latter we allow volumes to be taken which are a
little in excess of stated well capacity since the transfer wastage
volume (carry volume) should be a deliberate overestimate once a well is
nearly empty.

We therefore choose to silently correct autoallocation volumes to be at
the exact maximum when they are within a tolerance (currently the well
residual volume) of the max value, and generate an error if they exceed
this.

Test Plan:

antha-tester; manual inspection; additional unit tests (subsequently)

Reviewers:

Haydn King

Subscribers:

Projects:

Antha-language

* Regression tests for two cases of volume error

Summary:

These tests ensure that cases where a multichannel transfer is split up
into several multi-and-or-single channel transfers it does not transfer
incorrect volumes by including wells which should not be included

Test Plan:

These tests have been compared to master, in which they fail.
Antha-tester and manual inspection are also included

Reviewers:

Haydn King

Subscribers:

Projects:

Antha-language

* Clearer error message for regression test

Summary:

In the second case it was not sufficiently clear what the error was,
this is now rectified

Test Plan:

Reviewers:

Subscribers:

Projects:

* Revise comment about handling of zero transfers

Summary:

A hopefully clearer explanation of why we need to remove
zeroes when checking how much to multichannel

Test Plan:

visual inspection

Reviewers:

Haydn king

Subscribers:

Projects:

Antha language