lh.Plan() refactor: Cosmetic refactoring (#959)
the function microArch/scheduler/liquidhandling/liquidhandler.Plan() is the core part of our instruction generation, yet it's a nest of horror and a real pain point for development and on boarding.

I've had a crack at a refactor before, but ended up biting off more than I could chew in one go, but at least learnt what the chunks of work to do should be. This is part one, where I attempt to:

use camelCase consistently, at least for function calls
hang functions off the objects on which they operate as much as possible
-- sadly, this meant moving some assertions into wtype since LHInstruction currently lives there
push some complexity down
improve some code comments
This change should not affect overall behaviour of the system in any way, just make it easier on the eyes.