Add volume checks to lh simulator
Summary:

This branch includes a 'logical simulator', which compares
the declared outputs to the results of simulating the instructions
against an idealised liquid handler.

Most useful for testing instruction optimisations and catching edge
cases in the planning process, which makes several assumptions about the
volumes transferred

Test Plan:

Mostly passes with antha-tester, however some bundles only work if carry
volumes are set to 0. This may be a good idea in general, however it
implies some of the simulator logic needs checking to fix this