[ANTHA-2624] Log to disk.
Use kitlog's SwapLogger, make it possible to dynamically and safely change the writers a logger is going to, and document the behaviour.
Use this to ensure that both workflow and transpilation logs go to disk appropriately as soon as possible. Be pretty robust about syncing and saving the files too.