Make ShapeTypeID a struct, harder to inadvertently define a broken one (#994)
Previously, the shape type id was typedefd as a string. It was easy to inadvertently define a new object with a random shapetype (e.g. "trapezoid"), which then caused a crash if we attempted to access the IsRound method.
Now both the name and roundedness are required up-front, so this should no longer be an issue.