[ANTHA-2642] Support URLs for loading files.
We have a few concerns here: yes, if we get a simple http or https url, we can download it. We download it to the outdir, and then we need to be careful around the caching.
Came to the conclusing that caching with the index as whatever the user provides is the safest thing rather than messing about with FromSlash/ToSlash too early on.
We always make sure we cache everything (currently), which means that a read of a file that was written by a previous element is guaranteed to find content in the cache.
Which means that reads that miss the cache and are not http(s) urls can be enforced to be reads from the indir.