Improve "insufficient source volume" errors (#1005)
"Insufficient source volume" errors are notorious for being difficult to debug, and often consume a great deal of core developers time unnecessarily. This is because the error message produced contains little information and is difficult to interpret.
This change improves the quality of the error messages and simplifies the "GetSourcesFor" mechanism.