ANTHA-2689: Show correct transfer volumes in actions-summary (#1015)
Previously, if antha attempted to aspirate from a well that didn't have enough volume, the physical simulator reduced the aspirated volume to the maximum possible and raised a warning (which the users wouldn't see from the UI). This causes confusion as the new mix graph then sees the wrong volume being transferred.

Aspirating from empty or nearly empty wells can happen due to ANTHA-2704, where an accumulation of carry volumes consumes more volume than expected. In most cases, this is unlikely to be an issue in physical execution, since the carry volume is deliberately conservative and a full 0.5ul is unlikely to be spirited away on each transfer.

Fixing the underlying issue is likely to be quite a bit more work, see ANTHA-2704 for a full summary.

This proposed piecemeal step changes the physical simulator so that it:

no longer reduces transferred volumes so mix-graph works as expected
raises a (non-fatal) warning if a transfer invades the carry volume
raises an error (fatal) if a transfer exhausts the total volume in a well
Thus the user is insulated from trivial errors by the residual volume, but in more serious cases they are given an error and forced to work around the issue.
See ANTHA-2689 for more.