ANTHA-2690: Respect user Liquid names in actions summary (#1032)
Sometimes after issuing a mix instruction, elements will rename output liquids using liquid.SetName(...). Previously this was not respected in the generated actions.json, this change addresses that.

This is done by keeping track of the last reference to each well within a transferblock and updating the name of the liquid as given by the instruction - similarly to how this is already done for the FinalProperties object within liquidhandler.Plan, but for each transferblock.