Make file manager robust against "empty" file descriptors.
The reason for this is that it's basically unavoidable: the UI can always provide a value to a file parameter that is missing a path, whether we like it or not.
Because the parameters at the json level are totally arbitrary (json.RawMessage) there's no hope of doing validation early - we have to wait until workflow runtime.
So actually detecting this in the filemanager when it comes to ReadAll is both reasonable and robust.