Write the forgotten merge function for Meta.
Turns out that I'd forgotten to write a merge function for Meta which had the effect that the laboratory was never being passed the Name, which then shows up in the test output (elements).
Merging is ... a bit of a challenge given we have no schema for Meta. The chosen compromise is to just unmarshal as far as map[string]json.RawMessage and then do some byte comparison on matching keys.
This has necessitated the addition of a Meta.Set(key,value) function because otherwise the call site would have to have responsibility for serialisation itself, which would be nuts.
This all works, as evidenced by test output workflows:
```
{
  "SchemaVersion": "2.0",
  "WorkflowId": "hzA74iel7O",
  "SimulationId": "hzA74iel7O-dMlJouj1cuS",
  "Meta": {
    "Name": "DOE Media Optimisation Example",
    "SimulatorVersion": "unknown"
  },
  ...
```