[AC-34] Debitrot tests for microArch/scheduler/liquidhandling
This is a large change.
Many of the tests are not really unit tests; they are integration tests, and so rely on public apis from elsewhere in the codebase in order to do things like setup, or indeed execution.
Due to Go not allowing import cycles, several of the tests on their own were creating cycles, and so had to be moved to their own packages to break the cycle.
This in turn has meant that some of the APIs they were testing have had to be modified slightly to make them public, but this is no great hardship.

In order to aid breaking the import cycles, there have been other movements, particularly around the organisation of high and low-level instructions, which is very much beneficial anyway - everything was previously just dumped in effects/