Enable code coverage for elements_test
Because we have code that calls 'go test', this is a bit harder than the antha-test.sh version, but not too bad.
Basically just need to run 'go list' appropriately and pass the results through to the 'go test' commands.

Turns out using -f '{{ join .Deps "\n" }}' was wrong - it's sufficient just to list all the packages we can find. Because we're using go mod, provided we're in the right directory to run this, it's really straight forward stuff.

We currently generate two cover.out files - one for the compileAllElements test (which is really the 2nd part of that - the go test step within there to hit all the go tests); and one for the test workflows.
We may need to write a small amount of code to merge these cover files together but this is hardly the end of the world.
Interestingly, the test workflows certainly do not guarantee that every element is transpiled. Consequently the compileAllElements test is really the canonical source of truth for all element packages.