Build outdir into generated tests.
It turns out that because both elements_test and testlab were defining the same global flag, this was upsetting go during tests.
So the flag in testlab has gone: it was only used by the generated tests (workflows in elements repos), and there's nothing to be gained by attempting to be mobile: fact is that we bake in the indir anyway, so baking in the outdir doesn't make things worse, and it also means we can use the name of the workflow in the outdir which makes lining everything up much much easier.