ANTHA-2761 - support non-contiguous channels (#1043)
Occasionally the planner will attempt to multichannel using e.g. channels 0 and 2 on independent platforms like the Hamilton. This is possible, except that several places lower down in the instruction tree and the summary generation assume that channels are contiguous. This change removes this assumtion.

See: https://synthace.atlassian.net/browse/ANTHA-2761