[AC-65] Allow parts of the schema to be null.
This matches up with where we have omitEmpty in the workflow struct, and so this must match up so that workflows that we write out can be read back in again and be valid.

I'm annoyed by the change in case of Read in the go-generate - it suggests we have a different version of go-bindata installed. Oh well.