ANTHA-2791: Always build summaries, fix a few ins.Multi (#1053)
Previously on the command line, layout and actions summaries were only generated if the flags --layoutSummary or --mixSummary were provided respectively. This has led to a couple of missed bugs where summary generation failed bug wasn't detected until we got to the UI.

This change generates the summary data earlier so that any issues are caught immediately and fixes a few small bugs related to non-contiguous multi-channel that were highlighted by this change.