Changes so that Tecan EVO compiles
Summary:
(1) in liquidhandler.go I need to make sure that ignorePhysicalSimulation is set to true. We don't want to debug the physical simulation as it seems to be too low level
(2) in compositerobotinstructions I hardoced override to true for model evo because it doesn't make sense that we need to mix with volumes larger than the tips. The liquid policy regarding mix volumes needs to be reviewed anyway so this just is a workaround