Add comments clarifying differences between tipbox and plate handling
Summary
We handle plates and tipboxes differently: the former get their entire config through the front-end (as one of the options), the latter don't have frontend representations, and have to work with references only. It seems obvious an improved data model is required. In absence of such improved model I tried to make some notes clarifying the difference between tipboxes and plates