Adding partnumber to TipBoxes
Summary:
In order to match "Antha" representation of tips with Tecan EVO, I need the part number of each tip. This also makes tips and plates more similar. In order to do so I had to add partnumbers on the library elements. I set them to "" ( empty string) as we are about to review this model anyway.

Testing:
Ran my construct Assembly workflow